requirements.txt
fitz==0.0.1.dev2
openai==1.59.7
Pillow==11.1.0
opencv-python==4.10.0.84
numpy==2.2.1
matplotlib==3.10.0
torch==2.5.1
transformers==4.48.0
streamlit==1.41.1
pytesseract==0.3.13
scikit-image==0.25.0
pdf2image==1.17.0
pdfminer.six==20240706